+++ rpcs3/Emu/Cell/Modules/cellCamera.cpp
@@ -2,6 +2,7 @@
 #include "cellCamera.h"
 
 #include "Emu/System.h"
+#include "Emu/system_config.h"
 #include "Emu/Cell/PPUModule.h"
 #include "Emu/Cell/lv2/sys_event.h"
 #include "Emu/IdManager.h"
