+++ rpcs3/Emu/RSX/GL/GLHelpers.h
@@ -11,7 +11,6 @@
 #include "../Common/TextureUtils.h"
 #include "../Program/GLSLTypes.h"
 
-#include "Emu/system_config.h"
 #include "Utilities/mutex.h"
 #include "Utilities/geometry.h"
 #include "Utilities/File.h"
