+++ rpcs3/Emu/RSX/GL/GLOverlays.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "Emu/system_config_types.h"
 #include "util/types.hpp"
 #include "../Common/simple_array.hpp"
 #include "../Overlays/overlays.h"
