+++ rpcs3/Emu/RSX/Overlays/HomeMenu/overlay_home_menu_settings.cpp
@@ -1,6 +1,7 @@
 #include "stdafx.h"
 #include "overlay_home_menu_settings.h"
 #include "overlay_home_menu_components.h"
+#include "Emu/system_config.h"
 
 // TODO: Localization of the setting names
 // TODO: Localization of the dropdown values
